% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_r2anki_flashcard.R
\name{addin_r2anki_flashcard}
\alias{addin_r2anki_flashcard}
\title{Insert a new emtpy flashcard.}
\usage{
addin_r2anki_flashcard()
}
\description{
Call this function as an addin from RStudio to insert a new flashcard at the cursor position. \strong{Protipp}: Bind this function to a keyboard shortcut via Tools > Modify Keyboard Shortcuts. (Search for "addin" or "r2anki flashcard" to find the right functioncall.)
}
