% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css4anki.R
\name{css4anki}
\alias{css4anki}
\title{Get css code to style Anki cards with your preferred theme and code highlighting.}
\usage{
css4anki(highlight, theme)
}
\description{
This code then needs to be put in the css-style-section of your r2anki-note-type. You can create multiple note types in Anki and choose the one you like best to display your cards.
}
