% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_card_ids.R
\name{get_max_cardid}
\alias{get_max_cardid}
\title{Parse a file for cardids and return the highest number.}
\usage{
get_max_cardid(file, pattern = "^<!-- cardid: .* -->$")
}
\arguments{
\item{file}{Path to file to be parsed. Usually an RMarkdown-document.}

\item{pattern}{Regexp-pattern to define cardids.}
}
\description{
This is useful to find a starting-value to start counting for new card-ids.
}
