% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images2anki.R
\name{images2anki}
\alias{images2anki}
\title{Copy images from the \code{sourcefile} to the Anki \code{collection.media} folder}
\usage{
images2anki(rmd_sourcefile, user = "User 1",
  path_collections.media_folder = anki_dir(user), delete_directory = FALSE)
}
\arguments{
\item{rmd_sourcefile}{Path to the rmd-template}

\item{path_collections.media_folder}{Path to the Anki-template. Should be similar to /home/usrname/.local/share/Anki2/User 1/collection.media/"}

\item{delete_directory}{Should the image files be deleted after the copying?}
}
\description{
You will have to call this function, when you have images in your anki cards. When you see empty plots in your Anki-Cards you should copy the images over to the anki-collections.media-folder.
}
\seealso{
https://apps.ankiweb.net/docs/manual.html#files
}
