% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_anki.R
\name{initialize_anki}
\alias{initialize_anki}
\title{Open an apgk-file in Anki Which Contains Template Cards for appropriate r2anki-styling}
\usage{
initialize_anki()
}
\description{
Please import initialization file manually, if this function fails to find the *.apkg-file.
}
