% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{initialize_counter}
\alias{initialize_counter}
\title{Create an \code{.increase_cardid}-function, with a prespecified startpoint}
\usage{
initialize_counter(i = 0)
}
\description{
Create an \code{.increase_cardid}-function, with a prespecified startpoint
}
