% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_r2anki_flashcard.R
\name{insert_r2anki_flashcard}
\alias{insert_r2anki_flashcard}
\title{Insert a new emtpy flashcard.}
\usage{
insert_r2anki_flashcard()
}
\description{
Call this function as an addin to insert a new flashcard at the cursor position.
}
