% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_anki.R
\name{install_anki}
\alias{install_anki}
\title{Go to the Web and find yourself the up-to-date Anki-version for your OS.}
\usage{
install_anki()
}
\description{
This function will be substituted by a more automatic approach.
}
