% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2anki.R
\name{r2anki}
\alias{r2anki}
\title{General Function to enable the typical r2anki-workflow}
\usage{
r2anki(rmd_sourcefile, invoke_anki = FALSE, ...)
}
\arguments{
\item{invoke_anki}{Should Anki be started and the tsv imported? Defaults to \code{FALSE}}

\item{path_collections.media_folder}{Path to the Anki-template. Should be similar to \code{"/home/usrname/.local/share/Anki2/User 1/collection.media/"}}
}
\description{
General Function to enable the typical r2anki-workflow
}
