% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2tsv.R
\name{rmd2tsv}
\alias{rmd2tsv}
\title{Parse an Rmd-\code{r2anki}-\strong{template} to tsv (for Import into Anki)}
\usage{
rmd2tsv(rmd_sourcefile, ...)
}
\arguments{
\item{...}{Pass \code{quiet = TRUE} to \code{rmarkdown::render()} to suppress messages from pandoc comand line.}

\item{sourcefile}{path to \code{r2Anki}-Template}
}
\description{
The goal of \code{rmd2csv} is to compile an r2nki-Flashcard-Template. The output file is a tab-separed file. The name of the input-sourcefile should be uniqe across the collection, because the images in the Anki-cards will be named after the sourcefile and need to be unique within the Anki-database.
}
\examples{
## Define a Markdown-Template (read in from pkg-examples)
}
\references{
https://en.wikipedia.org/wiki/Spaced_repetition
}
\seealso{
\link{addin_r2anki_flashcard}
}
\author{
Henning Bumann, \email{h_b@posteo.de}
}
