% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shorten_documentname}
\alias{shorten_documentname}
\title{Shorten the name of the currently active file in the RStudio-Editor to a specified lenght.}
\usage{
shorten_documentname(stop = 12)
}
\arguments{
\item{stop}{Define length of the filename (to be used as a prefix for the cardid).}
}
\description{
Shorten the name of the currently active file in the RStudio-Editor to a specified lenght.
}
