% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv2anki.R
\name{tsv2anki}
\alias{tsv2anki}
\title{Start Anki (on Linux) and import a *.tsv-file to your Collections}
\usage{
tsv2anki(tsv_file)
}
\description{
Start Anki (on Linux) and import a *.tsv-file to your Collections
}
