% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse2csv.R
\name{parse2csv}
\alias{parse2csv}
\title{Parse an Rmd-\code{r2anki}-\strong{Template} to csv (for Import into Anki).}
\usage{
parse2csv(srcfile)
}
\arguments{
\item{sourcefile}{path to \code{r2Anki}-Template}
}
\description{
The goal of \code{parse2csv} is to compile an Anki-Rmd-Template.
}
\examples{
## Define a Markdown-Template (read in from pkg-examples)
}
\references{
https://en.wikipedia.org/wiki/Spaced_repetition
}
\seealso{
\link{addin_r2anki_flashcard}
}
\author{
Henning Bumann, \email{h_b@posteo.de}
}
