% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2anki.R
\name{r2anki}
\alias{r2anki}
\title{Main Function to Export r2anki-notebooks to Anki.}
\usage{
r2anki(rmd_sourcefile, invoke_anki = FALSE, init = FALSE, ...)
}
\arguments{
\item{rmd_sourcefile}{Path to R Markdown document to be parsed and exported to Anki.}

\item{invoke_anki}{Should Anki be started and the tsv imported? Defaults to \code{FALSE}}
}
\description{
Main Function to Export r2anki-notebooks to Anki.
}
