% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2tsv.R
\name{rmd2tsv}
\alias{rmd2tsv}
\title{Parse an Rmd-\code{r2anki}-\strong{Template} to csv (for Import into Anki).}
\usage{
rmd2tsv(rmd_sourcefile)
}
\arguments{
\item{sourcefile}{path to \code{r2Anki}-Template}
}
\description{
The goal of \code{parse2csv} is to compile an Anki-Rmd-Template. The name of the source-files should be uniqe across the collection, because the images will be named after the sourcefile and need to be unique within the anki-database.
}
\examples{
## Define a Markdown-Template (read in from pkg-examples)
}
\references{
https://en.wikipedia.org/wiki/Spaced_repetition
}
\seealso{
\link{addin_r2anki_flashcard}
}
\author{
Henning Bumann, \email{h_b@posteo.de}
}
